+++ glue/app/qt/application.py
@@ -907,8 +907,11 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
 
         # include file filter twice, so it shows up in Dialog
         outfile, file_filter = compat.getsavefilename(
-            parent=self, filters=("Glue Session (*.glu);; "
-                                  "Glue Session including data (*.glu)"))
+            parent=self, basedir=getattr(self, '_last_session_name', 'session.glu'),
+            filters=("Glue Session with absolute paths to data(*.glu);; "
+                     "Glue Session with relative paths to data(*.glu);; "
+                     "Glue Session including data (*.glu)"),
+            selectedfilter=getattr(self, '_last_session_filter', ''))
 
         # This indicates that the user cancelled
         if not outfile:
@@ -918,9 +921,13 @@ class GlueApplication(Application, QtWidgets.QMainWindow):
         if '.' not in outfile:
             outfile += '.glu'
 
+        self._last_session_name = outfile
+        self._last_session_filter = file_filter
+
         with set_cursor_cm(Qt.WaitCursor):
-            self.save_session(
-                outfile, include_data="including data" in file_filter)
+            self.save_session(outfile,
+                              include_data="including data" in file_filter,
+                              absolute_paths="absolute" in file_filter)
 
     @messagebox_on_error("Failed to export session")
     def _choose_export_session(self, saver, checker, outmode):